+++ router-tests/testenv/testenv.go
@@ -120,8 +120,9 @@ type Config struct {
 	DisableParentBasedSampler          bool
 	TLSConfig                          *core.TlsConfig
 	TraceExporter                      trace.SpanExporter
-	OtelAttributes                     []config.CustomAttribute
-	OtelResourceAttributes             []config.CustomStaticAttribute
+	CustomMetricAttributes             []config.CustomAttribute
+	CustomTelemetryAttributes          []config.CustomAttribute
+	CustomResourceAttributes           []config.CustomStaticAttribute
 	MetricReader                       metric.Reader
 	PrometheusRegistry                 *prometheus.Registry
 	ShutdownDelay                      time.Duration
@@ -700,8 +701,7 @@ func configureRouter(listenerAddr string, testConfig *Config, routerConfig *node
 
 		c := core.TraceConfigFromTelemetry(&config.Telemetry{
 			ServiceName:        "cosmo-router",
-			Attributes:         testConfig.OtelAttributes,
-			ResourceAttributes: testConfig.OtelResourceAttributes,
+			ResourceAttributes: testConfig.CustomResourceAttributes,
 			Tracing: config.Tracing{
 				Enabled:               true,
 				SamplingRate:          1,
@@ -715,7 +715,13 @@ func configureRouter(listenerAddr string, testConfig *Config, routerConfig *node
 
 		c.TestMemoryExporter = testConfig.TraceExporter
 
-		routerOpts = append(routerOpts, core.WithTracing(c))
+		routerOpts = append(routerOpts,
+			core.WithTracing(c),
+		)
+	}
+
+	if testConfig.CustomTelemetryAttributes != nil {
+		routerOpts = append(routerOpts, core.WithTelemetryAttributes(testConfig.CustomTelemetryAttributes))
 	}
 
 	var prometheusConfig rmetric.PrometheusConfig
@@ -736,10 +742,10 @@ func configureRouter(listenerAddr string, testConfig *Config, routerConfig *node
 	if testConfig.MetricReader != nil {
 		c := core.MetricConfigFromTelemetry(&config.Telemetry{
 			ServiceName:        "cosmo-router",
-			Attributes:         testConfig.OtelAttributes,
-			ResourceAttributes: testConfig.OtelResourceAttributes,
+			ResourceAttributes: testConfig.CustomResourceAttributes,
 			Tracing:            config.Tracing{},
 			Metrics: config.Metrics{
+				Attributes: testConfig.CustomMetricAttributes,
 				Prometheus: config.Prometheus{
 					Enabled: true,
 				},
@@ -1192,6 +1198,17 @@ func (e *Environment) GraphQLRequestURL() string {
 		urlStr = u.String()
 	}
 	return urlStr
+}
+
+func (e *Environment) MakeGraphQLMultipartRequest(method string, body io.Reader) *http.Request {
+	req, err := http.NewRequest(method, e.GraphQLRequestURL(), body)
+	require.NoError(e.t, err)
+
+	req.Header.Set("Content-Type", "application/json")
+	req.Header.Set("Accept", "multipart/mixed;subscriptionSpec=\"1.0\", application/json")
+	req.Header.Set("Connection", "keep-alive")
+
+	return req
 }
 
 func (e *Environment) GraphQLSubscriptionURL() string {