+++ router/core/operation_processor.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"crypto/sha256"
 	"fmt"
-	"github.com/wundergraph/graphql-go-tools/v2/pkg/middleware/operation_complexity"
 	"hash"
 	"io"
 	"net/http"
@@ -20,18 +19,17 @@ import (
 	"github.com/pkg/errors"
 	"github.com/tidwall/sjson"
 	fastjson "github.com/wundergraph/astjson"
+	"github.com/wundergraph/cosmo/router/internal/persistedoperation"
+	"github.com/wundergraph/cosmo/router/internal/unsafebytes"
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/ast"
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/astnormalization"
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/astparser"
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/astprinter"
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/astvalidation"
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/engine/datasource/httpclient"
+	"github.com/wundergraph/graphql-go-tools/v2/pkg/middleware/operation_complexity"
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/operationreport"
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/variablesvalidation"
-	"go.opentelemetry.io/otel/attribute"
-
-	"github.com/wundergraph/cosmo/router/internal/persistedoperation"
-	"github.com/wundergraph/cosmo/router/internal/unsafebytes"
 )
 
 var (
@@ -93,6 +91,7 @@ type OperationProcessorOptions struct {
 	QueryDepthCache                *ristretto.Cache[uint64, int]
 	OperationHashCache             *ristretto.Cache[uint64, string]
 	ParseKitPoolSize               int
+	IntrospectionEnabled           bool
 }
 
 // OperationProcessor provides shared resources to the parseKit and OperationKit.
@@ -104,6 +103,7 @@ type OperationProcessor struct {
 	operationCache           *OperationCache
 	parseKits                map[int]*parseKit
 	parseKitSemaphore        chan int
+	introspectionEnabled     bool
 }
 
 // parseKit is a helper struct to parse, normalize and validate operations
@@ -144,6 +144,7 @@ type OperationKit struct {
 	operationProcessor       *OperationProcessor
 	kit                      *parseKit
 	parsedOperation          *ParsedOperation
+	introspectionEnabled     bool
 }
 
 type GraphQLRequest struct {
@@ -171,6 +172,7 @@ func NewOperationKit(processor *OperationProcessor) *OperationKit {
 		operationDefinitionRef: -1,
 		cache:                  processor.operationCache,
 		parsedOperation:        &ParsedOperation{},
+		introspectionEnabled:   processor.introspectionEnabled,
 	}
 }
 
@@ -342,7 +344,7 @@ func (o *OperationKit) ComputeOperationSha256() error {
 
 // FetchPersistedOperation fetches the persisted operation from the cache or the client. If the operation is fetched from the cache it returns true.
 // UnmarshalOperationFromBody or UnmarshalOperationFromURL must be called before calling this method.
-func (o *OperationKit) FetchPersistedOperation(ctx context.Context, clientInfo *ClientInfo, commonTraceAttributes []attribute.KeyValue) (bool, error) {
+func (o *OperationKit) FetchPersistedOperation(ctx context.Context, clientInfo *ClientInfo) (bool, error) {
 	if o.operationProcessor.persistedOperationClient == nil {
 		return false, &httpGraphqlError{
 			message:    "could not resolve persisted query, feature is not configured",
@@ -359,13 +361,78 @@ func (o *OperationKit) FetchPersistedOperation(ctx context.Context, clientInfo *
 	if fromCache {
 		return true, nil
 	}
-	persistedOperationData, err := o.operationProcessor.persistedOperationClient.PersistedOperation(ctx, clientInfo.Name, o.parsedOperation.GraphQLRequestExtensions.PersistedQuery.Sha256Hash, commonTraceAttributes)
+
+	persistedOperationData, err := o.operationProcessor.persistedOperationClient.PersistedOperation(ctx, clientInfo.Name, o.parsedOperation.GraphQLRequestExtensions.PersistedQuery.Sha256Hash)
 	if err != nil {
 		return false, err
 	}
 	// it's important to make a copy of the persisted operation data, because it's used in the cache
 	// we might modify it later, so we don't want to modify the cached data
 	o.parsedOperation.Request.Query = string(persistedOperationData)
+
+	return false, nil
+}
+
+const (
+	schemaIntrospectionFieldName = "__schema"
+	typeIntrospectionFieldName   = "__type"
+)
+
+func (o *OperationKit) isIntrospectionQuery() (result bool, err error) {
+	var operationDefinitionRef = ast.InvalidRef
+	var possibleOperationDefinitionRefs = make([]int, 0)
+
+	for i := 0; i < len(o.kit.doc.RootNodes); i++ {
+		if o.kit.doc.RootNodes[i].Kind == ast.NodeKindOperationDefinition {
+			possibleOperationDefinitionRefs = append(possibleOperationDefinitionRefs, o.kit.doc.RootNodes[i].Ref)
+		}
+	}
+
+	if len(possibleOperationDefinitionRefs) == 0 {
+		return
+	} else if len(possibleOperationDefinitionRefs) == 1 {
+		operationDefinitionRef = possibleOperationDefinitionRefs[0]
+	} else {
+		for i := 0; i < len(possibleOperationDefinitionRefs); i++ {
+			ref := possibleOperationDefinitionRefs[i]
+			name := o.kit.doc.OperationDefinitionNameString(ref)
+
+			if o.parsedOperation.Request.OperationName == name {
+				operationDefinitionRef = ref
+				break
+			}
+		}
+	}
+
+	if operationDefinitionRef == ast.InvalidRef {
+		return
+	}
+
+	operationDef := o.kit.doc.OperationDefinitions[operationDefinitionRef]
+	if operationDef.OperationType != ast.OperationTypeQuery {
+		return
+	}
+	if !operationDef.HasSelections {
+		return
+	}
+
+	selectionSet := o.kit.doc.SelectionSets[operationDef.SelectionSet]
+	if len(selectionSet.SelectionRefs) == 0 {
+		return
+	}
+
+	for i := 0; i < len(selectionSet.SelectionRefs); i++ {
+		selection := o.kit.doc.Selections[selectionSet.SelectionRefs[i]]
+		if selection.Kind != ast.SelectionKindField {
+			continue
+		}
+
+		fieldName := o.kit.doc.FieldNameUnsafeString(selection.Ref)
+		switch fieldName {
+		case schemaIntrospectionFieldName, typeIntrospectionFieldName:
+			return true, nil
+		}
+	}
 
 	return false, nil
 }
@@ -392,6 +459,24 @@ func (o *OperationKit) Parse() error {
 	if report.HasErrors() {
 		return &reportError{
 			report: report,
+		}
+	}
+
+	if !o.introspectionEnabled {
+		isIntrospection, err := o.isIntrospectionQuery()
+
+		if err != nil {
+			return &httpGraphqlError{
+				message:    "could not determine if operation was an introspection query",
+				statusCode: http.StatusOK,
+			}
+		}
+
+		if isIntrospection {
+			return &httpGraphqlError{
+				message:    "GraphQL introspection is disabled by Cosmo Router, but the query contained __schema or __type. To enable introspection, set introspection_enabled: true in the Router configuration",
+				statusCode: http.StatusOK,
+			}
 		}
 	}
 
@@ -874,6 +959,7 @@ func NewOperationProcessor(opts OperationProcessorOptions) *OperationProcessor {
 		persistedOperationClient: opts.PersistedOperationClient,
 		parseKits:                make(map[int]*parseKit, opts.ParseKitPoolSize),
 		parseKitSemaphore:        make(chan int, opts.ParseKitPoolSize),
+		introspectionEnabled:     opts.IntrospectionEnabled,
 	}
 	for i := 0; i < opts.ParseKitPoolSize; i++ {
 		processor.parseKitSemaphore <- i