+++ router/core/router.go
@@ -16,18 +16,14 @@ import (
 	"connectrpc.com/connect"
 	"github.com/wundergraph/cosmo/router/pkg/watcher"
 
-	"github.com/wundergraph/cosmo/router/internal/persistedoperation"
-	"github.com/wundergraph/cosmo/router/internal/persistedoperation/cdn"
-	"github.com/wundergraph/cosmo/router/internal/persistedoperation/s3"
-	"github.com/wundergraph/cosmo/router/pkg/execution_config"
-	"github.com/wundergraph/cosmo/router/pkg/routerconfig"
-	configCDNProvider "github.com/wundergraph/cosmo/router/pkg/routerconfig/cdn"
-	configs3Provider "github.com/wundergraph/cosmo/router/pkg/routerconfig/s3"
-
 	"github.com/nats-io/nuid"
 	"github.com/redis/go-redis/v9"
 	"github.com/wundergraph/cosmo/router/internal/docker"
 	"github.com/wundergraph/cosmo/router/internal/graphiql"
+	"github.com/wundergraph/cosmo/router/internal/persistedoperation"
+	"github.com/wundergraph/cosmo/router/internal/persistedoperation/cdn"
+	"github.com/wundergraph/cosmo/router/internal/persistedoperation/s3"
+	"github.com/wundergraph/cosmo/router/pkg/execution_config"
 	"go.uber.org/atomic"
 
 	"github.com/mitchellh/mapstructure"
@@ -193,32 +189,23 @@ type (
 		accessLogsConfig          *AccessLogsConfig
 		// If connecting to localhost inside Docker fails, fallback to the docker internal address for the host
 		localhostFallbackInsideDocker bool
-
-		tlsServerConfig *tls.Config
-		tlsConfig       *TlsConfig
-
+		tlsServerConfig               *tls.Config
+		tlsConfig                     *TlsConfig
+		telemetryAttributes           []config.CustomAttribute
 		// Poller
-		configPoller configpoller.ConfigPoller
-		selfRegister selfregister.SelfRegister
-
-		registrationInfo *nodev1.RegistrationInfo
-
-		securityConfiguration config.SecurityConfiguration
-
-		customModules []Module
-
+		configPoller                 configpoller.ConfigPoller
+		selfRegister                 selfregister.SelfRegister
+		registrationInfo             *nodev1.RegistrationInfo
+		securityConfiguration        config.SecurityConfiguration
+		customModules                []Module
 		engineExecutionConfiguration config.EngineExecutionConfiguration
 		// should be removed once the users have migrated to the new overrides config
 		overrideRoutingURLConfiguration config.OverrideRoutingURLConfiguration
 		// the new overrides config
-		overrides config.OverridesConfiguration
-
-		authorization *config.AuthorizationConfiguration
-
-		rateLimit *config.RateLimitConfiguration
-
-		webSocketConfiguration *config.WebSocketConfiguration
-
+		overrides                config.OverridesConfiguration
+		authorization            *config.AuthorizationConfiguration
+		rateLimit                *config.RateLimitConfiguration
+		webSocketConfiguration   *config.WebSocketConfiguration
 		subgraphErrorPropagation config.SubgraphErrorPropagationConfiguration
 		clientHeader             config.ClientHeader
 	}
@@ -838,8 +825,7 @@ func (r *Router) buildClients() error {
 		}
 
 		c, err := cdn.NewClient(provider.URL, r.graphApiToken, cdn.Options{
-			Logger:        r.logger,
-			TraceProvider: r.tracerProvider,
+			Logger: r.logger,
 		})
 		if err != nil {
 			return err
@@ -874,8 +860,7 @@ func (r *Router) buildClients() error {
 		}
 
 		c, err := cdn.NewClient(r.cdnConfig.URL, r.graphApiToken, cdn.Options{
-			Logger:        r.logger,
-			TraceProvider: r.tracerProvider,
+			Logger: r.logger,
 		})
 		if err != nil {
 			return err
@@ -906,87 +891,12 @@ func (r *Router) buildClients() error {
 		r.persistedOperationClient = c
 	}
 
-	var rClient routerconfig.Client
-
-	// Poller is only initialized when a config poller is configured and the router is not started with a static config
-	if r.staticExecutionConfig == nil && r.routerConfigPollerConfig != nil && r.configPoller == nil {
-
-		if provider, ok := cdnProviders[r.routerConfigPollerConfig.Storage.ProviderID]; ok {
-
-			if r.graphApiToken == "" {
-				return errors.New(
-					"graph token is required to fetch execution config from CDN. " +
-						"Alternatively, configure a custom storage provider or specify a static execution config",
-				)
-			}
-
-			c, err := configCDNProvider.NewClient(
-				provider.URL,
-				r.graphApiToken,
-				&configCDNProvider.Options{
-					Logger:       r.logger,
-					SignatureKey: r.routerConfigPollerConfig.GraphSignKey,
-				})
-			if err != nil {
-				return err
-			}
-			rClient = c
-
-			r.logger.Info("Polling for execution config updates from CDN in the background",
-				zap.String("provider_id", provider.ID),
-				zap.String("interval", r.routerConfigPollerConfig.PollInterval.String()),
-			)
-		} else if provider, ok := s3Providers[r.routerConfigPollerConfig.Storage.ProviderID]; ok {
-
-			c, err := configs3Provider.NewClient(provider.Endpoint, &configs3Provider.ClientOptions{
-				AccessKeyID:     provider.AccessKey,
-				SecretAccessKey: provider.SecretKey,
-				BucketName:      provider.Bucket,
-				Region:          provider.Region,
-				ObjectPath:      r.routerConfigPollerConfig.Storage.ObjectPath,
-				Secure:          provider.Secure,
-			})
-			if err != nil {
-				return err
-			}
-			rClient = c
-
-			r.logger.Info("Polling for execution config updates from S3 storage in the background",
-				zap.String("provider_id", provider.ID),
-				zap.String("interval", r.routerConfigPollerConfig.PollInterval.String()),
-			)
-		} else {
-			if r.routerConfigPollerConfig.Storage.ProviderID != "" {
-				return fmt.Errorf("unknown storage provider id '%s' for execution config", r.routerConfigPollerConfig.Storage.ProviderID)
-			}
-
-			if r.graphApiToken == "" {
-				return errors.New(
-					"graph token is required to fetch execution config from CDN. " +
-						"Alternatively, configure a custom storage provider or specify a static execution config",
-				)
-			}
-
-			c, err := configCDNProvider.NewClient(r.cdnConfig.URL, r.graphApiToken, &configCDNProvider.Options{
-				Logger:       r.logger,
-				SignatureKey: r.routerConfigPollerConfig.GraphSignKey,
-			})
-			if err != nil {
-				return err
-			}
-			rClient = c
-
-			r.logger.Info("Polling for execution config updates from Cosmo CDN in the background",
-				zap.String("interval", r.routerConfigPollerConfig.PollInterval.String()),
-			)
-		}
-
-		r.configPoller = configpoller.New(r.graphApiToken,
-			configpoller.WithLogger(r.logger),
-			configpoller.WithPollInterval(r.routerConfigPollerConfig.PollInterval),
-			configpoller.WithClient(rClient),
-		)
-
+	configPoller, err := InitializeConfigPoller(r, cdnProviders, s3Providers)
+	if err != nil {
+		return err
+	}
+	if configPoller != nil {
+		r.configPoller = *configPoller
 	}
 
 	return nil
@@ -1659,6 +1569,12 @@ func WithTLSConfig(cfg *TlsConfig) Option {
 	}
 }
 
+func WithTelemetryAttributes(attributes []config.CustomAttribute) Option {
+	return func(r *Router) {
+		r.telemetryAttributes = attributes
+	}
+}
+
 func WithConfigPollerConfig(cfg *RouterConfigPollerConfig) Option {
 	return func(r *Router) {
 		r.routerConfigPollerConfig = cfg
@@ -1676,6 +1592,7 @@ func WithApolloCompatibilityFlagsConfig(cfg config.ApolloCompatibilityFlags) Opt
 		if cfg.EnableAll {
 			cfg.ValueCompletion.Enabled = true
 			cfg.TruncateFloats.Enabled = true
+			cfg.SuppressFetchErrors.Enabled = true
 		}
 		r.apolloCompatibilityFlags = cfg
 	}
@@ -1766,18 +1683,35 @@ func TraceConfigFromTelemetry(cfg *config.Telemetry) *rtrace.Config {
 		Sampler:            cfg.Tracing.SamplingRate,
 		ParentBasedSampler: cfg.Tracing.ParentBasedSampler,
 		WithNewRoot:        cfg.Tracing.WithNewRoot,
+		Attributes:         nil,
 		ExportGraphQLVariables: rtrace.ExportGraphQLVariables{
 			Enabled: cfg.Tracing.ExportGraphQLVariables,
 		},
-		SpanAttributesMapper: buildAttributesMapper(cfg.Attributes),
-		ResourceAttributes:   buildResourceAttributes(cfg.ResourceAttributes),
-		Exporters:            exporters,
-		Propagators:          propagators,
-		ResponseTraceHeader:  cfg.Tracing.ResponseTraceHeader,
+		ResourceAttributes:  buildResourceAttributes(cfg.ResourceAttributes),
+		Exporters:           exporters,
+		Propagators:         propagators,
+		ResponseTraceHeader: cfg.Tracing.ResponseTraceHeader,
 	}
 }
 
-func buildAttributesMapper(attributes []config.CustomAttribute) func(req *http.Request) []attribute.KeyValue {
+// buildAttributesMap returns a map of custom attributes to quickly check if a field is used in the custom attributes.
+func buildAttributesMap(attributes []config.CustomAttribute) map[string]string {
+	result := make(map[string]string)
+	for _, attr := range attributes {
+		if attr.ValueFrom != nil && attr.ValueFrom.ContextField != "" {
+			result[attr.ValueFrom.ContextField] = attr.Key
+		}
+	}
+	return result
+}
+
+// buildHeaderAttributesMapper returns a function that maps custom attributes to the request headers.
+func buildHeaderAttributesMapper(attributes []config.CustomAttribute) func(req *http.Request) []attribute.KeyValue {
+
+	if len(attributes) == 0 {
+		return nil
+	}
+
 	return func(req *http.Request) []attribute.KeyValue {
 		var result []attribute.KeyValue
 
@@ -1826,7 +1760,7 @@ func MetricConfigFromTelemetry(cfg *config.Telemetry) *rmetric.Config {
 	return &rmetric.Config{
 		Name:               cfg.ServiceName,
 		Version:            Version,
-		AttributesMapper:   buildAttributesMapper(cfg.Attributes),
+		Attributes:         cfg.Metrics.Attributes,
 		ResourceAttributes: buildResourceAttributes(cfg.ResourceAttributes),
 		OpenTelemetry: rmetric.OpenTelemetry{
 			Enabled:       cfg.Metrics.OTLP.Enabled,